Permit defaults to be any type in a union field
Strictly, the Avro specification mandates that the default of a
union-typed field must always match the first type in the union. In
practice, this mandate is frequently violated by schemas that are in
production use today. To be pragmatically useful, this change amends
Avroc's validation steps to allow a default value to match any type in
a union field.